fix: validate files in CreateChapter
Fix the security exploit: DOM text reinterpreted as HTML
This happened when the CreateChapter allow user to upload image and display a preview of it on the frontend without any file validation.
Many validation such as file size, file name along with error handling and fallback image have been implemented to fix this.